+++ module/flash.c
@@ -117,12 +117,15 @@ void flash_write(uint8_t preset_no, scene_state_t *scene,
 
 void flash_read(uint8_t preset_no, scene_state_t *scene,
                 char (*text)[SCENE_TEXT_LINES][SCENE_TEXT_CHARS],
-                uint8_t init_grid, uint8_t init_i2c_op_address) {
+                uint8_t init_pattern, uint8_t init_grid,
+                uint8_t init_i2c_op_address) {
     memcpy(ss_scripts_ptr(scene), &f.scenes[preset_no].scripts,
            // Exclude size of TEMP script as above
            ss_scripts_size() - sizeof(scene_script_t));
-    memcpy(ss_patterns_ptr(scene), &f.scenes[preset_no].patterns,
-           ss_patterns_size());
+    if (init_pattern) {
+        memcpy(ss_patterns_ptr(scene), &f.scenes[preset_no].patterns,
+               ss_patterns_size());
+    }
     if (init_grid) {
         memcpy(&grid_data, &f.scenes[preset_no].grid_data, sizeof(grid_data_t));
         unpack_grid(scene);