+++ src/ops/controlflow.h
@@ -20,6 +20,7 @@ extern const tele_op_t op_SYM_DOLLAR_POL;
 extern const tele_op_t op_KILL;
 extern const tele_op_t op_SCENE;
 extern const tele_op_t op_SCENE_G;
+extern const tele_op_t op_SCENE_P;
 extern const tele_op_t op_BREAK;
 extern const tele_op_t op_BRK;
 extern const tele_op_t op_SYNC;